Applied changes from the review
