+++ aspnetcore-webapi/UsingResultPatternInNETWebAPI/UsingResultPatternInNETWebAPI/Errors/RecordNotFoundError.cs
@@ -1,0 +1,9 @@
+﻿namespace UsingResultPatternInNETWebAPI.v5.Errors;
+
+public class RecordNotFoundError : Error
+{
+    public RecordNotFoundError(string message)
+        : base(message)
+    {
+    }
+}
