+++ aspnetcore-webapi/UsingResultPatternInNETWebAPI/UsingResultPatternInNETWebAPI/Errors/ValidationError.cs
@@ -1,0 +1,9 @@
+﻿namespace UsingResultPatternInNETWebAPI.v5.Errors;
+
+public class ValidationError : Error
+{
+    public ValidationError(string message)
+        : base(message)
+    {
+    }
+}
