+++ aspnetcore-webapi/UsingResultPatternInNETWebAPI/UsingResultPatternInNETWebAPI/Exceptions/RecordNotFoundException.cs
@@ -1,0 +1,9 @@
+﻿namespace UsingResultPatternInNETWebAPI.v3.Exceptions;
+
+public class RecordNotFoundException : Exception
+{
+    public RecordNotFoundException(string message)
+        : base(message)
+    {
+    }
+}
