+++ aspnetcore-webapi/UsingResultPatternInNETWebAPI/UsingResultPatternInNETWebAPI/Exceptions/ValidationException.cs
@@ -1,0 +1,9 @@
+﻿namespace UsingResultPatternInNETWebAPI.v3.Exceptions;
+
+public class ValidationException : Exception
+{
+    public ValidationException(string message)
+        : base(message)
+    {
+    }
+}
