--- aspnetcore-webapi/UsingResultPatternInNETWebAPI/UsingResultPatternInNETWebAPI/Repositories/IContactsRepository.cs
@@ -1,9 +1,0 @@
-﻿namespace UsingResultPatternInNETWebAPI.Repositories;
-
-public interface IContactsRepository
-{
-    IEnumerable<Contact> GetAll();
-    Contact? GetById(Guid id);
-    Contact Create(Contact contact);
-    bool ExistsByEmail(string email);
-}
