--- aspnetcore-webapi/UsingResultPatternInNETWebAPI/UsingResultPatternInNETWebAPI/v3/Exceptions/ApiNotFoundException.cs
@@ -1,9 +1,0 @@
-﻿namespace UsingResultPatternInNETWebAPI.v3.Exceptions;
-
-public class ApiNotFoundException : Exception
-{
-    public ApiNotFoundException(Guid id)
-        : base($"contact with id {id} was not found")
-    {
-    }
-}
