--- aspnetcore-webapi/UsingResultPatternInNETWebAPI/UsingResultPatternInNETWebAPI/v3/Exceptions/ApiValidationException.cs
@@ -1,9 +1,0 @@
-﻿namespace UsingResultPatternInNETWebAPI.v3.Exceptions;
-
-public class ApiValidationException : Exception
-{
-    public ApiValidationException(string message)
-        : base(message)
-    {
-    }
-}
