--- aspnetcore-webapi/UsingResultPatternInNETWebAPI/UsingResultPatternInNETWebAPI/v5/Errors/ApiNotFoundError.cs
@@ -1,9 +1,0 @@
-﻿namespace UsingResultPatternInNETWebAPI.v5.Errors;
-
-public class ApiNotFoundError : Error
-{
-    public ApiNotFoundError(Guid id)
-        : base($"contact with id {id} was not found")
-    {
-    }
-}
