--- aspnetcore-webapi/UsingResultPatternInNETWebAPI/UsingResultPatternInNETWebAPI/v5/Errors/ApiValidationError.cs
@@ -1,9 +1,0 @@
-﻿namespace UsingResultPatternInNETWebAPI.v5.Errors;
-
-public class ApiValidationError : Error
-{
-    public ApiValidationError(string message)
-        : base(message)
-    {
-    }
-}
