--- aspnetcore-webapi/UsingResultPatternInNETWebAPI/UsingResultPatternInNETWebAPI/v5/Services/FluentResultsContactsService.cs
@@ -1,46 +1,0 @@
-﻿using UsingResultPatternInNETWebAPI.v5.Errors;
-
-namespace UsingResultPatternInNETWebAPI.v5.Services;
-
-public class FluentResultsContactsService
-{
-    private readonly IContactsRepository _contactsRepository;
-
-    public FluentResultsContactsService(IContactsRepository contactsRepository)
-    {
-        _contactsRepository = contactsRepository;
-    }
-
-    public Result<IEnumerable<ContactDto>> GetAll()
-    {
-        var contactDtos = _contactsRepository
-            .GetAll()
-            .Select(c => new ContactDto(c.Id, c.Email));
-
-        return Result.Ok(contactDtos);
-    }
-
-    public Result<ContactDto> GetById(Guid id)
-    {
-        var contact = _contactsRepository.GetById(id);
-
-        if (contact is null)
-        {
-            return new ApiNotFoundError(id);
-        }
-
-        return Result.Ok(new ContactDto(contact.Id, contact.Email));
-    }
-
-    public Result<ContactDto> Create(CreateContactDto contact)
-    {
-        if (_contactsRepository.ExistsByEmail(contact.Email))
-        {
-            return new ApiValidationError("contact with this email already exists");
-        }
-
-        var createdContact = _contactsRepository.Create(new Contact {Email = contact.Email});
-
-        return Result.Ok(new ContactDto(createdContact.Id, createdContact.Email));
-    }
-}
